refactor: make file name more proper
